feature: 디렉터리 변경 및 임시 any 처리, ServerSideProps api 적용
