+++ next/components/video-chat/ChatSection/Chat/TextChat.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import * as S from './style';
 
-const TextChat = ({ nickname, content }) => {
+const TextChat = ({ nickname, content }: any) => {
   return (
     <S.ChatWrapper>
       <S.Writer>{nickname}</S.Writer>
