--- next/pages/video-chat.tsx
@@ -1,25 +1,0 @@
-import CustomHead from '@components/common/CustomHead';
-import Container from '@components/video-chat/Container';
-import CamSection from '@components/video-chat/CamSection';
-import ChatSection from '@components/video-chat/ChatSection';
-
-const pageProps = {
-  title: '화상채팅 - Mogakco',
-  description: 'Free online video chat for developers',
-  url: '', // TODO: 도메인 정해지면 url에 추가하기
-  locale: 'ko_KR',
-};
-
-const VideoChat = () => {
-  return (
-    <>
-      <CustomHead {...pageProps} />
-      <Container>
-        <CamSection />
-        <ChatSection />
-      </Container>
-    </>
-  );
-};
-
-export default VideoChat;
